Cleanup of setbuilder and starting to work on setbuilder. Starting from the 'evaluation' stage of a current set, for improvements and other things. New stat mods regarding defensive types, can remove defensive prop of battle items. Moves are now doubles so that their values can be altered finely. Replaced all floats to doubles. A few more randomizers in place

+++ IndymonProgram/MechanicsData/MovesAndAbilities.cs
@@ -83,8 +83,8 @@ namespace MechanicsData
         public string Name { get; set; } = "";
         public PokemonType Type { get; set; }
         public MoveCategory Category { get; set; }
-        public int Bp { get; set; }
-        public int Acc { get; set; }
+        public double Bp { get; set; }
+        public double Acc { get; set; }
         public HashSet<EffectFlag> Flags { get; set; } = new HashSet<EffectFlag>();
         public override string ToString()
         {
